requirements.txt
streamlit==0.73
matplotlib
pillow
numpy
https://github.com/google-coral/pycoral/releases/download/release-frogfish/tflite_runtime-2.5.0-cp36-cp36m-linux_x86_64.whl
wget
gspread